requirements.txt
opencv-python
mediapipe
pandas
numpy
customtkinter
pillow
